# Performance: Compressed chunks reads improval by 12.5% CPU Time