--- a/src/gorilla.c
+++ b/src/gorilla.c
@@ -114,22 +114,58 @@
 #define CMPR_L4 15
 #define CMPR_L5 32
 
+// The powers of 2 from 0 to 63
+static u_int64_t bittt[] = {
+    1ULL << 0,  1ULL << 1,  1ULL << 2,  1ULL << 3,  1ULL << 4,  1ULL << 5,  1ULL << 6,  1ULL << 7,
+    1ULL << 8,  1ULL << 9,  1ULL << 10, 1ULL << 11, 1ULL << 12, 1ULL << 13, 1ULL << 14, 1ULL << 15,
+    1ULL << 16, 1ULL << 17, 1ULL << 18, 1ULL << 19, 1ULL << 20, 1ULL << 21, 1ULL << 22, 1ULL << 23,
+    1ULL << 24, 1ULL << 25, 1ULL << 26, 1ULL << 27, 1ULL << 28, 1ULL << 29, 1ULL << 30, 1ULL << 31,
+    1ULL << 32, 1ULL << 33, 1ULL << 34, 1ULL << 35, 1ULL << 36, 1ULL << 37, 1ULL << 38, 1ULL << 39,
+    1ULL << 40, 1ULL << 41, 1ULL << 42, 1ULL << 43, 1ULL << 44, 1ULL << 45, 1ULL << 46, 1ULL << 47,
+    1ULL << 48, 1ULL << 49, 1ULL << 50, 1ULL << 51, 1ULL << 52, 1ULL << 53, 1ULL << 54, 1ULL << 55,
+    1ULL << 56, 1ULL << 57, 1ULL << 58, 1ULL << 59, 1ULL << 60, 1ULL << 61, 1ULL << 62, 1ULL << 63
+};
+
+static uint64_t bitmask[] = {
+    (1ULL << 0) - 1,  (1ULL << 1) - 1,  (1ULL << 2) - 1,  (1ULL << 3) - 1,  (1ULL << 4) - 1,
+    (1ULL << 5) - 1,  (1ULL << 6) - 1,  (1ULL << 7) - 1,  (1ULL << 8) - 1,  (1ULL << 9) - 1,
+    (1ULL << 10) - 1, (1ULL << 11) - 1, (1ULL << 12) - 1, (1ULL << 13) - 1, (1ULL << 14) - 1,
+    (1ULL << 15) - 1, (1ULL << 16) - 1, (1ULL << 17) - 1, (1ULL << 18) - 1, (1ULL << 19) - 1,
+    (1ULL << 20) - 1, (1ULL << 21) - 1, (1ULL << 22) - 1, (1ULL << 23) - 1, (1ULL << 24) - 1,
+    (1ULL << 25) - 1, (1ULL << 26) - 1, (1ULL << 27) - 1, (1ULL << 28) - 1, (1ULL << 29) - 1,
+    (1ULL << 30) - 1, (1ULL << 31) - 1, (1ULL << 32) - 1, (1ULL << 33) - 1, (1ULL << 34) - 1,
+    (1ULL << 35) - 1, (1ULL << 36) - 1, (1ULL << 37) - 1, (1ULL << 38) - 1, (1ULL << 39) - 1,
+    (1ULL << 40) - 1, (1ULL << 41) - 1, (1ULL << 42) - 1, (1ULL << 43) - 1, (1ULL << 44) - 1,
+    (1ULL << 45) - 1, (1ULL << 46) - 1, (1ULL << 47) - 1, (1ULL << 48) - 1, (1ULL << 49) - 1,
+    (1ULL << 50) - 1, (1ULL << 51) - 1, (1ULL << 52) - 1, (1ULL << 53) - 1, (1ULL << 54) - 1,
+    (1ULL << 55) - 1, (1ULL << 56) - 1, (1ULL << 57) - 1, (1ULL << 58) - 1, (1ULL << 59) - 1,
+    (1ULL << 60) - 1, (1ULL << 61) - 1, (1ULL << 62) - 1, (1ULL << 63) - 1,
+
+};
+
 // 2^bit
 static inline u_int64_t BIT(u_int64_t bit) {
     if (__builtin_expect(bit > 63, 0)) {
         return 0ULL;
     }
-    return (1ULL << bit);
+    return bittt[bit];
 }
 
 // the LSB `bits` turned on
 static inline u_int64_t MASK(u_int64_t bits) {
-    return BIT(bits) - 1;
+    if (__builtin_expect(bits > 63, 0)) {
+        return 0ULL - 1;
+    }
+    return bitmask[bits];
 }
 
+// Logic to check Least Significant Bit (LSB) of a number
 // Clear most significant bits from position `bits`
 static inline u_int64_t LSB(u_int64_t x, u_int64_t bits) {
-    return x & MASK(bits);
+    if (__builtin_expect(bits > 63, 0)) {
+        return x & (0ULL - 1);
+    }
+    return x & bitmask[bits];
 }
 
 /*
@@ -166,8 +202,7 @@ static binary_t int2bin(int64_t x, u_int8_t l) {
 
 // Converts `bin`, a binary of length `l` bits, into an int64
 static int64_t bin2int(binary_t bin, u_int8_t l) {
-    bool pos = !(bin & BIT(l - 1));
-    if (pos)
+    if (!(bin & BIT(l - 1)))
         return bin;
     // return (int64_t) (bin | ~MASK(l)); // sign extend `bin`
     return (int64_t)bin - BIT(l); // same but cheaper
@@ -185,7 +220,7 @@ static inline int64_t Bin_MinVal(u_int8_t nbits) {
 
 // `bit` is a global bit (can be out of scope of a single binary_t)
 
-static inline u_int8_t localbit(globalbit_t bit) {
+static inline u_int8_t localbit(const globalbit_t bit) {
     return bit % BINW;
 }
 
@@ -196,21 +231,17 @@ static bool Bin_InRange(int64_t x, u_int8_t nbits) {
     return x >= Bin_MinVal(nbits) && x <= Bin_MaxVal(nbits);
 }
 
-static inline binary_t *Bins_bitbin(u_int64_t *bins, globalbit_t bit) {
-    return &bins[bit / BINW];
+static inline bool Bins_bitoff(const u_int64_t *bins, globalbit_t bit) {
+    return !(bins[bit >> 6] & BIT(localbit(bit)));
 }
 
-static inline bool Bins_bitoff(u_int64_t *bins, globalbit_t bit) {
-    return !(bins[bit / BINW] & BIT(localbit(bit)));
-}
-
-static inline bool Bins_biton(u_int64_t *bins, globalbit_t bit) {
+static inline bool Bins_biton(const u_int64_t *bins, globalbit_t bit) {
     return !Bins_bitoff(bins, bit);
 }
 
 // Append `dataLen` bits from `data` into `bins` at bit position `bit`
 static void appendBits(binary_t *bins, globalbit_t *bit, binary_t data, u_int8_t dataLen) {
-    binary_t *bin_it = Bins_bitbin(bins, *bit);
+    binary_t *bin_it = &bins[(*bit) >> 6];
     localbit_t lbit = localbit(*bit);
     localbit_t available = BINW - lbit;
 
@@ -224,20 +255,16 @@ static void appendBits(binary_t *bins, globalbit_t *bit, binary_t data, u_int8_t
 }
 
 // Read `dataLen` bits from `bins` at position `bit`
-static binary_t readBits(binary_t *bins, globalbit_t *bit, u_int8_t dataLen) {
-    binary_t *bin_it = Bins_bitbin(bins, *bit);
-    localbit_t lbit = localbit(*bit);
-    localbit_t available = BINW - lbit;
-
-    binary_t bin = 0;
+static inline binary_t readBits(const binary_t *bins,
+                                globalbit_t start_pos,
+                                const u_int8_t dataLen) {
+    const localbit_t lbit = localbit(start_pos);
+    const localbit_t available = BINW - lbit;
     if (available >= dataLen) {
-        bin = LSB(*bin_it >> lbit, dataLen);
-    } else {
-        u_int8_t left = dataLen - available;
-        bin = LSB(*bin_it >> lbit, available);
-        bin |= LSB(*++bin_it, left) << available;
+        return LSB(bins[start_pos >> 6] >> lbit, dataLen);
     }
-    *bit += dataLen;
+    binary_t bin = LSB(bins[start_pos >> 6] >> lbit, available);
+    bin |= LSB(bins[(start_pos >> 6) + 1], dataLen - available) << available;
     return bin;
 }
 
@@ -248,7 +275,9 @@ static bool isSpaceAvailable(CompressedChunk *chunk, u_int8_t size) {
 
 /***************************** APPEND ********************************/
 static ChunkResult appendInteger(CompressedChunk *chunk, timestamp_t timestamp) {
+#ifdef DEBUG
     assert(timestamp >= chunk->prevTimestamp);
+#endif
     timestamp_t curDelta = timestamp - chunk->prevTimestamp;
 
     union64bits doubleDelta;
@@ -328,10 +357,14 @@ static ChunkResult appendFloat(CompressedChunk *chunk, double value) {
 
     localbit_t prevLeading = chunk->prevLeading;
     localbit_t prevTrailing = chunk->prevTrailing;
+#ifdef DEBUG
     assert(leading + trailing <= BINW);
+#endif
     localbit_t blockSize = BINW - leading - trailing;
     u_int32_t expectedSize = DOUBLE_LEADING + DOUBLE_BLOCK_SIZE + blockSize;
-    assert(prevLeading + prevTrailing <= BINW);
+#ifdef DEBUG
+    assert(leading + trailing <= BINW);
+#endif
     localbit_t prevBlockInfoSize = BINW - prevLeading - prevTrailing;
     /*
      * First bit encodes whether previous block parameters can be used since
@@ -361,7 +394,9 @@ static ChunkResult appendFloat(CompressedChunk *chunk, double value) {
 }
 
 ChunkResult Compressed_Append(CompressedChunk *chunk, timestamp_t timestamp, double value) {
+#ifdef DEBUG
     assert(chunk);
+#endif
 
     if (chunk->count == 0) {
         chunk->baseValue.d = chunk->prevValue.d = value;
@@ -390,31 +425,35 @@ ChunkResult Compressed_Append(CompressedChunk *chunk, timestamp_t timestamp, dou
  * then decodes the value back to an int64 and calculate the original value
  * using `prevTS` and `prevDelta`.
  */
-static u_int64_t readInteger(Compressed_Iterator *iter) {
-    binary_t *bins = iter->chunk->data;
-    globalbit_t *bit = &iter->idx;
-
+static inline u_int64_t readInteger(Compressed_Iterator *iter, const uint64_t *bins) {
     int64_t dd = 0;
+    // control bit ‘0’
     // Read stored double delta value
-    if (Bins_bitoff(bins, (*bit)++)) {
+    if (Bins_bitoff(bins, iter->idx++)) {
         dd = 0;
-    } else if (Bins_bitoff(bins, (*bit)++)) {
-        dd = bin2int(readBits(bins, bit, CMPR_L1), CMPR_L1);
-    } else if (Bins_bitoff(bins, (*bit)++)) {
-        dd = bin2int(readBits(bins, bit, CMPR_L2), CMPR_L2);
-    } else if (Bins_bitoff(bins, (*bit)++)) {
-        dd = bin2int(readBits(bins, bit, CMPR_L3), CMPR_L3);
-    } else if (Bins_bitoff(bins, (*bit)++)) {
-        dd = bin2int(readBits(bins, bit, CMPR_L4), CMPR_L4);
-    } else if (Bins_bitoff(bins, (*bit)++)) {
-        dd = bin2int(readBits(bins, bit, CMPR_L5), CMPR_L5);
+    } else if (Bins_bitoff(bins, iter->idx++)) {
+        dd = bin2int(readBits(bins, iter->idx, CMPR_L1), CMPR_L1);
+        iter->idx += CMPR_L1;
+    } else if (Bins_bitoff(bins, iter->idx++)) {
+        dd = bin2int(readBits(bins, iter->idx, CMPR_L2), CMPR_L2);
+        iter->idx += CMPR_L2;
+    } else if (Bins_bitoff(bins, iter->idx++)) {
+        dd = bin2int(readBits(bins, iter->idx, CMPR_L3), CMPR_L3);
+        iter->idx += CMPR_L3;
+    } else if (Bins_bitoff(bins, iter->idx++)) {
+        dd = bin2int(readBits(bins, iter->idx, CMPR_L4), CMPR_L4);
+        iter->idx += CMPR_L4;
+    } else if (Bins_bitoff(bins, iter->idx++)) {
+        dd = bin2int(readBits(bins, iter->idx, CMPR_L5), CMPR_L5);
+        iter->idx += CMPR_L5;
     } else {
-        dd = readBits(bins, bit, 64);
+        dd = readBits(bins, iter->idx, 64);
+        iter->idx += 64;
     }
 
     // Update iterator
     iter->prevDelta += dd;
-    return iter->prevTS = iter->prevTS + iter->prevDelta;
+    return iter->prevTS += iter->prevDelta;
 }
 
 /*
@@ -427,51 +466,63 @@ static u_int64_t readInteger(Compressed_Iterator *iter) {
  *
  * Finally, the compressed representation of the value is decoded.
  */
-static double readFloat(Compressed_Iterator *iter) {
-    binary_t xorValue;
-    union64bits rv;
-
+static inline double readFloat(Compressed_Iterator *iter, const uint64_t *data) {
     // Check if value was changed
-    if (Bins_bitoff(iter->chunk->data, iter->idx++)) {
+    // control bit ‘0’ (case a)
+    if (Bins_bitoff(data, iter->idx++)) {
         return iter->prevValue.d;
     }
+    binary_t xorValue;
 
-    // Check if previous block information was used
-    bool usePreviousBlockInfo = Bins_bitoff(iter->chunk->data, iter->idx++);
-    if (usePreviousBlockInfo) {
-        assert(iter->prevLeading + iter->prevTrailing <= BINW);
-        u_int8_t prevBlockInfo = BINW - iter->prevLeading - iter->prevTrailing;
-        xorValue = readBits(iter->chunk->data, &iter->idx, prevBlockInfo);
-        xorValue <<= iter->prevTrailing;
+    // Check if we can use the previous block info
+    // meaning control bit number 2 is 1. i.e. control bits are ‘10’ (case b)
+    // there are at least as many leading zeros and as
+    // many trailing zeros as with the previous value
+    // use  the previous block  information and
+    // just read the meaningful XORed value
+    if (Bins_bitoff(data, iter->idx++)) {
+#ifdef DEBUG
+        assert(iter->leading + iter->trailing <= BINW);
+#endif
+        xorValue = readBits(data, iter->idx, iter->blocksize);
+        iter->idx += iter->blocksize;
+        xorValue <<= iter->trailing;
     } else {
-        binary_t leading = readBits(iter->chunk->data, &iter->idx, DOUBLE_LEADING);
-        binary_t blocksize =
-            readBits(iter->chunk->data, &iter->idx, DOUBLE_BLOCK_SIZE) + DOUBLE_BLOCK_ADJUST;
+        // Read the length of the number of leading zeros in the next 5 bits,
+        // then read the length of the meaningful XORed value in the next 6 bits.
+        // Finally read the meaningful bits of theXORed value
+        iter->leading = readBits(data, iter->idx, DOUBLE_LEADING);
+        iter->idx += DOUBLE_LEADING;
+        iter->blocksize = readBits(data, iter->idx, DOUBLE_BLOCK_SIZE) + DOUBLE_BLOCK_ADJUST;
+        iter->idx += DOUBLE_BLOCK_SIZE;
+#ifdef DEBUG
         assert(leading + blocksize <= BINW);
-        binary_t trailing = BINW - leading - blocksize;
-        xorValue = readBits(iter->chunk->data, &iter->idx, blocksize) << trailing;
-        iter->prevLeading = leading;
-        iter->prevTrailing = trailing;
+#endif
+        iter->trailing = BINW - iter->leading - iter->blocksize;
+        xorValue = readBits(data, iter->idx, iter->blocksize) << iter->trailing;
+        iter->idx += iter->blocksize;
     }
-
+    union64bits rv;
     rv.u = xorValue ^ iter->prevValue.u;
     return iter->prevValue.d = rv.d;
 }
 
 ChunkResult Compressed_ReadNext(Compressed_Iterator *iter, timestamp_t *timestamp, double *value) {
+#ifdef DEBUG
     assert(iter);
     assert(iter->chunk);
-
+#endif
     if (iter->count >= iter->chunk->count)
         return CR_END;
-
-    if (iter->count == 0) { // First sample
+    // First sample
+    if (__builtin_expect(iter->count == 0, 0)) {
         *timestamp = iter->chunk->baseTimestamp;
         *value = iter->chunk->baseValue.d;
-    } else {
-        *timestamp = iter->prevTS = readInteger(iter);
-        *value = iter->prevValue.d = readFloat(iter);
+        iter->count++;
+        return CR_OK;
     }
+    *timestamp = readInteger(iter, iter->chunk->data);
+    *value = readFloat(iter, iter->chunk->data);
     iter->count++;
     return CR_OK;
 }