--- a/tests/flow/test_ts_madd.py
+++ b/tests/flow/test_ts_madd.py
@@ -74,3 +74,25 @@ def test_partial_madd():
         assert len(r.execute_command('ts.range', 'test_key1', "-", "+")) == 2
         assert len(r.execute_command('ts.range', 'test_key2', "-", "+")) == 2
         assert len(r.execute_command('ts.range', 'test_key3', "-", "+")) == 2
+
+
+def test_extensive_ts_madd():
+    with Env(decodeResponses=True).getConnection() as r:
+        r.execute_command("ts.create", 'test_key1')
+        r.execute_command("ts.create", 'test_key2')
+        pos = 1
+        lines = []
+        float_lines = []
+        with open("lemire_canada.txt","r") as file:
+            lines = file.readlines()
+        for line in lines:
+            float_v = float(line.strip())
+            res = r.execute_command("ts.madd", 'test_key1', pos, float_v, 'test_key2', pos, float_v)
+            assert res == [pos,pos]
+            pos=pos+1
+            float_lines.append(float_v)
+        returned_floats = r.execute_command('ts.range', 'test_key1', "-", "+")
+        assert len(returned_floats) == len(float_lines)
+        for pos,datapoint in enumerate(returned_floats,start=1):
+            assert pos == datapoint[0]
+            assert float_lines[pos-1] == float(datapoint[1])