This PR optimizes the READs of the compressed chunks, namely by:
- reduce as much as possible the updates by reference passing
- reuse precomputed temp values
- precalculate all possible data and have on two auxiliary tables
- move away from condition check the readInteger and readFloat meaning we’re helping reduce mispredicted branches
- inline and simplify, readFloat and readInteger

## Hotpots by self CPU utilization that were positively impacted
Function | CPU Time difference | Master SELF CPU time | Optimized SELF CPU time | Source File
-- | -- | -- | -- | --
readBits | 9.60% | 15.90% | 6.30% | gorilla.c
Bins_bitoff | 1.40% | 2.50% | 1.10% | gorilla.c
Compressed_ChunkIteratorGetNext | 0.60% | 6.40% | 5.80% | compressed_chunk.c
Bins_bitbin | 0.80% | 0.90% | 0.10% | gorilla.c
localbit | 0.30% | 0.70% | 0.40% | gorilla.c
-- | -- | -- | -- | --
**Total** | **12.70%** | 26.40% | 13.70% | 


### Benchmark results

#### master branch on single-groupby-5-8-1 query
```
Summary:
  throughput summary: 571.46 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        1.716     1.656     1.711     1.791     2.207     2.863
```

#### compressed.improvements branch on single-groupby-5-8-1 query
```
Summary:
  throughput summary: 643.07 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
        1.522     1.456     1.511     1.599     2.047     2.783
```

---------------------

#### Steps to reproduce:

- Dataset:  tsbs devops use case, scale 100  `wget --no-check-certificate https://s3.amazonaws.com/benchmarks.redislabs/redistimeseries/tsbs/datasets/devops/functional/scale-100-redistimeseries_data.rdb`

- Benchmark command: `redis-benchmark -c 1 -l -n 20000 -e "TS.MRANGE" "1610665140000" "1610668740000" "AGGREGATION" "MAX" "60000" "WITHLABELS" "FILTER" "measurement=cpu" "fieldname=(usage_user,usage_system,usage_idle,usage_nice,usage_iowait)" "hostname=(host_9,host_3,host_5,host_1,host_7,host_2,host_8,host_4)" "GROUPBY" "fieldname" "REDUCE" "MAX"`